feat: 添加PlayerPoints支持并重构经济相关代码
添加对PlayerPoints经济插件的支持，允许用户选择Vault或PlayerPoints作为经济提供者
重构经济相关代码，将金额类型从double改为int以适配PlayerPoints
修改配置文件结构，增加经济提供者选项并调整金额格式
添加PlayerPointsHook类处理PlayerPoints相关操作
更新VaultHook类以支持int类型金额

+++ src/main/java/com/oolonghoo/woomni/command/NicknameCommand.java
@@ -108,7 +108,7 @@ public class NicknameCommand implements CommandExecutor, TabCompleter {
         }
         
         if (!isSettingOthers && module.isEconomyEnabled()) {
-            double cost = module.getSetCost(target);
+            int cost = module.getSetCost(target);
             if (cost > 0) {
                 if (!module.canAfford(target)) {
                     msg.send(sender, "nickname.not-enough-money", "cost", module.formatCost(cost));
@@ -122,7 +122,7 @@ public class NicknameCommand implements CommandExecutor, TabCompleter {
         String processedNick = processNickname(nickname, target, module);
         
         if (!isSettingOthers && module.isEconomyEnabled()) {
-            double cost = module.getSetCost(target);
+            int cost = module.getSetCost(target);
             if (cost > 0 && !module.chargePlayer(target)) {
                 msg.send(sender, "nickname.payment-failed");
                 return true;
